+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-2.2.0
+3.0.0
 -----
 - Detect aggregate report email attachments by file content rather than
 file extension
+- In an aggregate report's `org_name` is a FQDN, the base is used
 - Add option to select the IMAP folder where reports are stored
-- Update CLI help
+- Add options to send data to Elasticsearch
 
 2.1.2
 -----